Add --reuse-orig option to make-deb-src-pkg.sh
git-svn-id: svn://svn.gna.org/svn/plash/trunk@428 e2ebe5e9-4d0c-0410-8560-844d39661dae

+++ make-deb-src-pkg.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 
 # Create Debian source package.
-# Usage:  make-deb-src-pkg.sh <version> <debian-variant>...
+# Usage:  make-deb-src-pkg.sh [--reuse-orig] <version> <debian-variant>...
 # <debian-variant> is the name of the "debian" directory to use.
 #
 # eg. To make packages for all variants:
@@ -12,11 +12,20 @@
 # slightly different tarballs each time (i.e. with different
 # timestamps and file orderings).
 # This saves space.
+#
+# The --reuse-orig option allows reusing an existing .orig tarball,
+# which is taken from the current directory.
 
 set -e
 
+CREATE_ORIG=yes
+if [ "$1" = "--reuse-orig" ]; then
+  CREATE_ORIG=no
+  shift
+fi
+
 if [ $# -lt 2 ]; then
-  echo "Usage: $0 <version> <debian-variant>..."
+  echo "Usage: $0 [--reuse-orig] <version> <debian-variant>..."
   exit 1
 fi
 
@@ -42,12 +51,19 @@ rm -rv autom4te.cache
 rm -rv web-site/screenshot-*.png
 )
 
-# Create the .orig tarball without any debian directories in it
-echo "---- creating .orig tarball"
-cp -lr $DEST.first-orig $DEST.orig
-rm $DEST.orig/debian
-rm -r $DEST.orig/debian-*
-tar -czf $OUT/plash_$VERSION.orig.tar.gz -C $OUT plash-$VERSION.orig
+if [ "$CREATE_ORIG" = yes ]; then
+  # Create the .orig tarball without any debian directories in it
+  echo "---- creating .orig tarball"
+  cp -lr $DEST.first-orig $DEST.orig
+  rm $DEST.orig/debian
+  rm -r $DEST.orig/debian-*
+  tar -czf $OUT/plash_$VERSION.orig.tar.gz -C $OUT plash-$VERSION.orig
+else
+  echo "---- reusing .orig tarball"
+  cp -l plash_$VERSION.orig.tar.gz $OUT/
+  # Unpack it for the benefit of "dpkg-source -ss" below
+  tar -xf $OUT/plash_$VERSION.orig.tar.gz -C $OUT
+fi
 
 # For each variant of debian, create source package
 for VARIANT in $@; do
@@ -56,7 +72,9 @@ for VARIANT in $@; do
   cp -lr $OUT/plash-$VERSION.orig $OUT/$VARIANT/
   cp -lr $OUT/plash_$VERSION.orig.tar.gz $OUT/$VARIANT/
   
-  cp -lr $DEST.orig $OUT/$VARIANT/plash-$VERSION
+  cp -lr $DEST.first-orig $OUT/$VARIANT/plash-$VERSION
+  rm $OUT/$VARIANT/plash-$VERSION/debian
+  rm -r $OUT/$VARIANT/plash-$VERSION/debian-*
   cp -lr $DEST.first-orig/$VARIANT $OUT/$VARIANT/plash-$VERSION/debian
   # The -ss option tells dpkg-source that both the directory
   # and the tarfile are available.  It uses the directory for
